Split Service Bus client tests into method-specific files
